Update c3_secure_database_access.mdx
@UlisesGascon - This section is a carbon copy of the first paragraph. There may be a paste error and you missed what you should have pasted under Secure Queries.

### SQL Injection occurs when untrusted user input is dynamically added to a SQL query in an
insecure manner, often via basic string concatenation. SQL Injection is one of the most
dangerous application security risks. SQL Injection is easy to exploit and could lead to the
entire database being stolen, wiped, or modified. The application can even be used to run
dangerous commands against the operating system hosting your database, thereby giving an
attacker a foothold on your network.
In order to mitigate SQL injection, untrusted input should be prevented from being
interpreted as part of a SQL command. The best way to do this is with the programming
technique known as ‘Query Parameterization’. This defense should be applied to SQL, OQL, as
well as stored procedure construction.